adding a new file structure
